fix multi-platform warnings
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@45159 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/algo/structure/struct_util/su_sequence_set.cpp
@@ -177,7 +177,7 @@ static void StringFromStdaa(const vector < char >& vec, string *str)
 
     str->resize(vec.size());
     for (unsigned int i=0; i<vec.size(); ++i)
-        str->at(i) = stdaaMap[vec[i]];
+        str->at(i) = stdaaMap[(unsigned int) vec[i]];
 }
 
 Sequence::Sequence(ncbi::objects::CBioseq& bioseq) :
